Move FacilityBadge.ts to libraries
