+++ components/alert/AlertMap.vue
@@ -154,7 +154,7 @@ import zoneNameFilter from '~/filters/ZoneName'
 import { FacilityType } from '~/constants/FacilityType'
 import { MapRegion } from '~/libraries/MapRegion'
 import { Faction } from '~/constants/Faction'
-import { FacilityBadge } from '~/interfaces/FacilityBadge'
+import { FacilityBadge } from '~/libraries/FacilityBadge'
 import factionShortName from '~/filters/FactionShortName'
 import factionCircleEmoji from '~/filters/FactionCircleEmoji'
 import { InstanceOutfitAggregateResponseInterface } from '~/interfaces/aggregates/instance/InstanceOutfitAggregateResponseInterface'
